refactor: Make NonBlockingSocket a trait to allow alternate socket types
Instead of creating the socket itself, the sessions now get a socket
trait object passed into the constructor.

start_p2p_session et al are now responsible for creating the sockets. In
future commits, we may want to consider what kind of mechanism we want
to use for injecting other kinds of sockets, it could be done
conditionally based on feature flags in p2p_session, or perhaps we could
even allow foreign code to pass in an existing socket.
