+++ src/lib.rs
@@ -7,11 +7,12 @@
 
 //#![warn(clippy::all, clippy::pedantic, clippy::nursery, clippy::cargo)]
 
-use std::net::SocketAddr;
+use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 
 pub use error::GGRSError;
 pub use frame_info::{GameInput, GameState};
 pub use network::network_stats::NetworkStats;
+use network::non_blocking_socket::UdpNonBlockingSocket;
 pub use sessions::p2p_session::P2PSession;
 pub use sessions::p2p_spectator_session::P2PSpectatorSession;
 pub use sessions::sync_test_session::SyncTestSession;
@@ -30,9 +31,9 @@ pub(crate) mod sessions {
 pub(crate) mod network {
     pub(crate) mod compression;
     pub(crate) mod network_stats;
+    pub(crate) mod non_blocking_socket;
     pub(crate) mod udp_msg;
     pub(crate) mod udp_protocol;
-    pub(crate) mod udp_socket;
 }
 
 // #############
@@ -211,8 +212,13 @@ pub fn start_p2p_session(
             info: "Input size too big.".to_owned(),
         });
     }
-    P2PSession::new(num_players, input_size, local_port)
-        .map_err(|_| GGRSError::SocketCreationFailed)
+
+    // udp nonblocking socket creation
+    let addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), local_port); //TODO: IpV6?
+    let socket =
+        Box::new(UdpNonBlockingSocket::new(addr).map_err(|_| GGRSError::SocketCreationFailed)?);
+
+    Ok(P2PSession::new(num_players, input_size, socket))
 }
 
 /// Used to create a new `P2PSpectatorSession` for a spectator.
@@ -251,6 +257,16 @@ pub fn start_p2p_spectator_session(
             info: "Input size too big.".to_owned(),
         });
     }
-    P2PSpectatorSession::new(num_players, input_size, local_port, host_addr)
-        .map_err(|_| GGRSError::SocketCreationFailed)
+
+    // udp nonblocking socket creation
+    let addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), local_port); //TODO: IpV6?
+    let socket =
+        Box::new(UdpNonBlockingSocket::new(addr).map_err(|_| GGRSError::SocketCreationFailed)?);
+
+    Ok(P2PSpectatorSession::new(
+        num_players,
+        input_size,
+        socket,
+        host_addr,
+    ))
 }
