+++ src/network/non_blocking_socket/udp_socket.rs
@@ -1,16 +1,21 @@
+use std::{
+    io::ErrorKind,
+    net::{SocketAddr, ToSocketAddrs, UdpSocket},
+};
+
 use crate::network::udp_msg::UdpMessage;
-use std::io::ErrorKind;
-use std::net::{SocketAddr, ToSocketAddrs, UdpSocket};
+
+use super::NonBlockingSocket;
 
 const RECV_BUFFER_SIZE: usize = 4096;
 
 #[derive(Debug)]
-pub(crate) struct NonBlockingSocket {
+pub(crate) struct UdpNonBlockingSocket {
     socket: UdpSocket,
     buffer: [u8; RECV_BUFFER_SIZE],
 }
 
-impl NonBlockingSocket {
+impl UdpNonBlockingSocket {
     pub(crate) fn new<A: ToSocketAddrs>(addr: A) -> Result<Self, std::io::Error> {
         let socket = UdpSocket::bind(addr)?;
         socket.set_nonblocking(true)?;
@@ -19,13 +24,15 @@ impl NonBlockingSocket {
             buffer: [0; RECV_BUFFER_SIZE],
         })
     }
+}
 
-    pub(crate) fn send_to<A: ToSocketAddrs>(&self, msg: &UdpMessage, addr: A) {
+impl NonBlockingSocket for UdpNonBlockingSocket {
+    fn send_to(&self, msg: &UdpMessage, addr: SocketAddr) {
         let buf = bincode::serialize(&msg).unwrap();
         self.socket.send_to(&buf, addr).unwrap();
     }
 
-    pub(crate) fn receive_all_messages(&mut self) -> Vec<(SocketAddr, UdpMessage)> {
+    fn receive_all_messages(&mut self) -> Vec<(SocketAddr, UdpMessage)> {
         let mut received_messages = Vec::new();
         loop {
             match self.socket.recv_from(&mut self.buffer) {
